gpt: apparently key is a 32-bit integer
Or it is aligned or something. Eitherway, a byte is not enough despite
what the reverse engineered document says.

+++ src/musician-gpt-types.h
@@ -24,7 +24,7 @@
 G_BEGIN_DECLS
 
 typedef gint32 MusicianGptNote;
-typedef gint8  MusicianGptKey;
+typedef gint32 MusicianGptKey;
 
 typedef enum
 {
